Revert "Split catalog writer class into two classes (#97)"
This reverts commit e465538144b7df3280b6efa94852deddca00375f.
