+++ src/main/scala/services/lakehouse/base/CatalogWriter.scala
@@ -47,7 +47,12 @@ object S3CatalogFileIO extends S3CatalogFileIO:
   override val endpoint: String = scala.util.Properties.envOrElse("ARCANE_FRAMEWORK__S3_CATALOG_ENDPOINT", "")
   override val region: String = scala.util.Properties.envOrElse("ARCANE_FRAMEWORK__S3_CATALOG_REGION", "us-east-1")
 
-
+trait CatalogWriterBuilder[CatalogImpl, TableImpl, SchemaImpl]:
+  /**
+   * Initialize the catalog connection
+   * @return CatalogWriter instance ready to perform data operations
+   */
+  def initialize(): CatalogWriter[CatalogImpl, TableImpl, SchemaImpl]
   
 trait CatalogWriter[CatalogImpl, TableImpl, SchemaImpl] extends AutoCloseable:
   implicit val catalog: CatalogImpl
